+++ Jarvis.Framework.Shared/IdentitySupport/OfflineCounterService.cs
@@ -4,6 +4,7 @@ using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace Jarvis.Framework.Shared.IdentitySupport
 {
@@ -124,6 +125,21 @@ namespace Jarvis.Framework.Shared.IdentitySupport
             {
                 yield return new OfflineIdentity(serie, i);
             }
+        }
+
+        public Task<long> GetNextAsync(string serie)
+        {
+            return Task.FromResult(this.GetNext(serie));
+        }
+
+        public Task ForceNextIdAsync(string serie, long nextIdToReturn)
+        {
+            throw new NotImplementedException("Cannot force in offline.");
+        }
+
+        public Task<long> PeekNextAsync(string serie)
+        {
+            throw new NotImplementedException("In offline service it is not safe to know next id, we have not root counter.");
         }
     }
 }